# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Med Labs API Documentation",
        "description": "This is the documentation for the API",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "",
            "description": "This is the documentation for the API"
        }
    ],
    "paths": {
...
        "/api/v1/users": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Display all Users",
                "description": "Display a listing of the resource.",
                "operationId": "6404deedf4ff57b0ed2b004cb5ac1d7b",
                "responses": {
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Users"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Unauthenticated"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "UnprocessableEntity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnprocessableEntity"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFound"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ServerError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Create User",
                "description": "Store a newly created resource in storage.",
                "operationId": "ad2dee2e6c7a94acd72da2b02dd90958",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UserPostRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Unauthenticated"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "UnprocessableEntity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnprocessableEntity"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFound"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ServerError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "User"
                ],
                "summary": "Update User",
                "description": "Update the specified resource in storage.",
                "operationId": "ed8f38496e4aca41ecb3c47062e074e8",
                "parameters": [
                    {
                        "name": "name",
                        "in": null,
                        "description": "The name of the user",
                        "required": false,
                        "example": "John Doe"
                    },
                    {
                        "name": "email",
                        "in": null,
                        "description": "Email of the user",
                        "required": false,
                        "example": "john@example.com"
                    },
                    {
                        "name": "password",
                        "in": null,
                        "description": "The name of the User",
                        "required": false,
                        "example": "123456"
                    },
                    {
                        "name": "password_confirmation",
                        "in": null,
                        "description": "The name of the User",
                        "required": false,
                        "example": "123456"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Unauthenticated"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "UnprocessableEntity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnprocessableEntity"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFound"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ServerError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/user/{user_id}": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Display User",
                "description": "Display the specified resource.",
                "operationId": "b3788828365e755340e2ea85ccc12c3f",
                "responses": {
                    "200": {
                        "description": "Successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Unauthenticated"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "UnprocessableEntity",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnprocessableEntity"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFound"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ServerError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "Unauthenticated": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Unauthenticated"
                    },
                    "type": {
                        "type": "string",
                        "example": "AuthenticationException"
                    },
                    "status": {
                        "type": "integer",
                        "example": 401
                    }
                },
                "type": "object"
            },
            "NotFound": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Not Found"
                    },
                    "type": {
                        "type": "string",
                        "example": "NotFoundHttpException"
                    },
                    "status": {
                        "type": "integer",
                        "example": 404
                    }
                },
                "type": "object"
            },
            "UnprocessableEntity": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Unprocessable Content"
                    },
                    "type": {
                        "type": "string",
                        "example": "UnprocessableEntityException"
                    },
                    "status": {
                        "type": "integer",
                        "example": 422
                    }
                },
                "type": "object"
            },
            "ServerError": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "Internal Server Error"
                    },
                    "type": {
                        "type": "string",
                        "example": "ErrorException"
                    },
                    "status": {
                        "type": "integer",
                        "example": 500
                    }
                },
                "type": "object"
            },
...
            "UserPostRequest": {
                "required": [
                    "firstname",
                    "lastname",
                    "phone",
                    "email",
                    "password",
                    "password_confirmation"
                ],
                "properties": {
                    "firstname": {
                        "description": "First name of user",
                        "type": "string",
                        "example": "John"
                    },
                    "lastname": {
                        "description": "Last name of user",
                        "type": "string",
                        "example": "Doe"
                    },
                    "phone": {
                        "description": "Last name of user",
                        "type": "string",
                        "example": "1-984-993-4778"
                    },
                    "country": {
                        "description": "Last name of user",
                        "type": "string",
                        "example": "Philippines"
                    },
                    "state": {
                        "description": "State or province",
                        "type": "string",
                        "example": "Leyte"
                    },
                    "city": {
                        "description": "City or municipality",
                        "type": "string",
                        "example": "Tacloban"
                    },
                    "email": {
                        "description": "Unique email",
                        "type": "string",
                        "format": "email",
                        "example": "john@example.com"
                    },
                    "password": {
                        "type": "string",
                        "example": "123456"
                    },
                    "password_confirmation": {
                        "type": "string",
                        "example": "123456"
                    },
                    "owner_id": {
                        "description": "The owner (user_id) of this account",
                        "type": "integer",
                        "example": "1"
                    },
                    "role": {
                        "description": "Role",
                        "type": "integer",
                        "example": "1"
                    }
                },
                "type": "object"
            },
            "Users": {
                "properties": {
                    "users": {
                        "type": "array",
                        "items": {
                            "properties": {
                                "id": {
                                    "type": "integer",
                                    "example": 1
                                },
                                "firstname": {
                                    "type": "string",
                                    "example": "John"
                                },
                                "lastname": {
                                    "type": "string",
                                    "example": "Doe"
                                },
                                "email": {
                                    "type": "string",
                                    "example": "john@example.com"
                                },
                                "status": {
                                    "type": "string",
                                    "example": "ACTIVE"
                                },
                                "role": {
                                    "type": "string",
                                    "example": "TENANT"
                                },
                                "owner_id": {
                                    "type": "integer",
                                    "example": null
                                },
                                "phone": {
                                    "type": "string",
                                    "example": "123-456-6789"
                                },
                                "country": {
                                    "type": "string",
                                    "example": "Philippines"
                                },
                                "state": {
                                    "type": "string",
                                    "example": "Leyte"
                                },
                                "city": {
                                    "type": "string",
                                    "example": "Tacloban City"
                                },
                                "created_at": {
                                    "type": "string",
                                    "format": "date-time",
                                    "example": "2022-08-27T16:14:46.000000Z"
                                }
                            },
                            "type": "object"
                        }
                    },
                    "metadata": {
                        "properties": {
                            "count": {
                                "type": "integer",
                                "example": 1
                            },
                            "offset": {
                                "type": "integer",
                                "example": 0
                            },
                            "limit": {
                                "type": "string",
                                "example": 20
                            },
                            "sort": {
                                "type": "array",
                                "items": {},
                                "example": [
                                    "+id"
                                ]
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "User": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "firstname": {
                        "type": "string",
                        "example": "John"
                    },
                    "lastname": {
                        "type": "string",
                        "example": "Doe"
                    },
                    "email": {
                        "type": "string",
                        "example": "john@example.com"
                    },
                    "status": {
                        "type": "string",
                        "example": "ACTIVE"
                    },
                    "role": {
                        "type": "string",
                        "example": "TENANT"
                    },
                    "owner_id": {
                        "type": "integer",
                        "example": null
                    },
                    "phone": {
                        "type": "string",
                        "example": "123-456-6789"
                    },
                    "country": {
                        "type": "string",
                        "example": "Philippines"
                    },
                    "state": {
                        "type": "string",
                        "example": "Leyte"
                    },
                    "city": {
                        "type": "string",
                        "example": "Tacloban City"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2022-08-27T16:14:46.000000Z"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    }
}